requirements.txt
streamlit
google-generativeai
python-dotenv
plotly
requests